use version of harness that sends u64 like it used to

+++ prover/witness_generator/src/basic_circuits.rs
@@ -649,7 +649,7 @@ async fn generate_witness(
             },
             |a, b, c| {
                 sender
-                    .blocking_send(Message::RecursionQueue(a, b, c))
+                    .blocking_send(Message::RecursionQueue(a as u8, b, c))
                     .unwrap()
             },
         );
